Add files via upload
This C++ program demonstrates the use of pointers. It defines an integer variable num, initializes it to 2, and creates a pointer ptr that stores the address of num. The program then outputs the address of num, the address stored in the pointer, the value of num, and the value of num accessed via the pointer. This illustrates how pointers can be used to reference and manipulate variables in memory.