Add files via upload
This C++ program demonstrates exception handling with the out_of_range exception. The getElement() function attempts to access an element from an array, throwing an out_of_range exception if the index is invalid. In the main() function, the user is prompted to input an index, and the program tries to retrieve the corresponding array value. If the index is out of bounds, an error message is displayed.