Add files via upload
This C++ program demonstrates nested exception handling. The user is prompted to enter their age. The inner try block checks if the age is negative and throws an exception if so. If the exception is caught, it is rethrown to the outer catch block, where it is caught again and an appropriate error message is displayed. If the age is valid, it is printed.